coverage: embed untested cssutils in untested old renderer
This way when the renderer is removed the unused cssutils will be too.
