+++ src/utils/CSSUtils.ts
@@ -39,47 +39,6 @@ export function parseGuifont(guifont: string, defaults: any) {
         }, result as any);
 }
 
-// Takes a `guifont` declaration and returns that same font declaration but as
-// a bunch of CSS declarations.
-export function guifontToMultiDecl(guifont: string) {
-    const defaults: any = {};
-    defaults[fontFamily] = "Monospace";
-    defaults["font-size"] = "9pt";
-    return Object.entries(parseGuifont(guifont, defaults))
-        .map(([key, value]) => `${key}: ${value};\n`)
-        .join("\n");
-}
-
-// Takes an array of `guifont` declarations and returns them as a single CSS
-// declaration, using font-family for font fallback.
-export function guifontsToFontFamily(guifonts: string[]) {
-    const defaults: any = {};
-    defaults[fontFamily] = "Monospace";
-    defaults["font-size"] = "9pt";
-    const reducedGuifonts = guifonts
-        .slice()
-        .reverse()
-        .reduce((acc, cur) => parseGuifont(cur, acc), defaults);
-    return `font-family: ${reducedGuifonts[fontFamily]}; font-size: ${reducedGuifonts["font-size"]};`;
-}
-
-// Takes a string formatted according to the `guifont` spec and returns a CSS
-// declaration that matches it.
-export function guifontsToCSS(guifont: string) {
-    const guifonts = (guifont + ",")
-        .match(/.+?[^\\],/g) // split on non-escaped commas
-        .map(s => s.slice(0, -1)); // remove last comma of each font
-    if (guifonts.length > 1) {
-        // If there are multiple font declarations, we use a CSS declaration
-        // like this: `font-family: font-family1 font-size font-style
-        // font-weight, font-family2...`. This prevents us from setting
-        // size/bold/italics/underlnie/strikethrough but enables letting the
-        // browser fallback to other fonts if one can't be found.
-        return guifontsToFontFamily(guifonts);
-    }
-    return guifontToMultiDecl(guifonts[0]);
-}
-
 // Computes a unique selector for its argument.
 export function computeSelector(element: HTMLElement) {
     function uniqueSelector(e: HTMLElement): string {
@@ -111,15 +70,3 @@ export function toHexCss(n: number) {
     return "#" + (new Array(6 - str.length)).fill("0").join("") + str;
 }
 
-export function toHighlightClassName(n: number) {
-    return "nvim_highlight_" + n;
-}
-
-// Computes a CSS stylesheet that represents the HighlightArray
-export function toCss(highlights: HighlightArray) {
-    const bg = highlights[0].background;
-    const fg = highlights[0].foreground;
-    return highlights.reduce((css, elem, id) => css +
-        `.${toHighlightClassName(id)}{background: ${elem.background || bg};color:${elem.foreground || fg};font-style:${elem.italic ? "italic" : "normal"};font-weight:${elem.bold ? "bold" : "normal"};text-decoration-line:${(elem.undercurl || elem.underline) ? "underline" : (elem.strikethrough ? "line-through" : "none")};text-decoration-style:${elem.undercurl ? "wavy" : "solid"};}`
-        , "");
-}
